update markdoc configs
