+++ src/lib/layouts/DocsArticle.svelte
@@ -4,7 +4,7 @@
         href: string;
         step?: number;
         selected?: boolean;
-        level?: number;
+        level: number;
         children?: Array<{
             title: string;
             href: string;
