+++ src/markdoc/layouts/Post.svelte
@@ -128,12 +128,12 @@
                 <div class="border-smooth md:border-r md:pr-12 lg:col-span-9">
                     <PostMeta {authorData} {title} {timeToRead} {currentURL} {date} {description} />
                     {#if cover}
-                        <div class="web-media aspect-video">
+                        <div class="web-media my-8! aspect-video">
                             <Media class="block aspect-video object-cover" src={cover} />
                         </div>
                     {/if}
 
-                    <div class="web-article-content text-secondary mt-8 flex flex-col gap-8">
+                    <div class="web-article-content prose mt-8 flex flex-col gap-8">
                         {#if lastUpdated}
                             <span class="text-main-body last-updated-text font-medium">
                                 Updated:
