--- src/markdoc/nodes/Blockquote.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-    import { type Snippet } from 'svelte';
-    interface BlockquoteProps {
-        children: Snippet;
-    }
-
-    const { children }: BlockquoteProps = $props();
-</script>
-
-<blockquote>
-    <p class="leading-sm text-primary text-lg font-medium tracking-tight">
-        {@render children()}
-    </p>
-</blockquote>
